requirements.txt
beautifulsoup4==4.14.3
boto3==1.42.83
firecrawl_py==4.21.1
pandas==3.0.2
pdfplumber==0.11.9
Requests==2.33.1
